+++ examples/web-dashboard/src/main/scala/com/twitter/web/dashboard/views/UserView.scala
@@ -1,0 +1,8 @@
+package com.twitter.web.dashboard.views
+
+import com.twitter.finatra.response.Mustache
+
+@Mustache("user")
+case class UserView(
+  firstName: String,
+  lastName: String)